done PWM implementation

+++ steer_controller/firmware/mod_ems22a.cpp
@@ -37,7 +37,7 @@
 int ENCODER_ABS_CS  = 0x10;
 int ENCODER_ABS_CLK = 0x11;
 int ENCODER_ABS_DO  = 0x12;
-unsigned long  EMS22A_last_millis        = 0;
+uint32_t  EMS22A_last_millis        = 0;
 uint16_t       EMS22A_sampling_period_ms_tenths = 2000;
 bool           EMS22A_active             = false;
