+++ steer_controller/firmware/process_internal_adc.cpp
@@ -41,7 +41,7 @@
 // ADC reading subsystem:
 uint8_t        num_active_ADC_channels = 0;
 uint8_t        ADC_active_channels[MAX_ADC_CHANNELS] = {0,0,0,0,0,0,0,0};
-unsigned long  ADC_last_millis = 0;
+uint32_t  ADC_last_millis = 0;
 uint16_t       ADC_sampling_period_ms_tenths = 2000;
 
 void processADCs()
